requirements.txt
asgiref==3.8.1
bcrypt==4.3.0
Django==5.1.4
dnspython==2.7.0
email-validator==2.3.0
idna==3.10
mysql-connector-python==9.4.0
python-dotenv==1.1.1
sqlparse==0.5.3
